# src/app.ts
import { readFile } from 'node:fs/promises';

import express, { Request, Response, NextFunction } from 'express';
import cors from 'cors';
import multer from 'multer';
import * as OpenApiValidator from 'express-openapi-validator';
import { OpenAPIV3 } from 'express-openapi-validator/dist/framework/types.js';
import swaggerUi from 'swagger-ui-express';

import { config } from './config.js';
import { AuthenticatedUser, authenticate } from './auth.js';
import { createThumbnail, measure, probe } from './images.js';
import { renderHomePage, renderImagePage } from './page.js';
import {
    ID_PATTERN,
    ImageFormat,
    ImageMetadata,
    deleteImage,
    formatOf,
    exists,
    idOf,
    originalPath,
    readMetadata,
    linkFileAtomically,
    saveImage,
    thumbnailPath,
    urlOf,
    writeFileAtomically
} from './storage.js';
import apiDocs from './user-images.openapi.json';

/**
 * The routes are regular expressions rather than express paths, since an id has to be matched
 * exactly - anything else would let a request walk out of the storage directory.
 */
const IMAGE_ROUTE = new RegExp(`^/(${ID_PATTERN})\\.(jpg|png|webp)$`);
const METADATA_ROUTE = new RegExp(`^/(${ID_PATTERN})\\.json$`);
const PAGE_ROUTE = new RegExp(`^/(${ID_PATTERN})$`);

/** A picture at a given url can never change, so whoever holds it can keep it forever */
const IMMUTABLE_CACHE_CONTROL = 'public, max-age=31536000, immutable';

export const app = express();
app.use(cors());
app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(apiDocs));

// Whoever uploads is known before the upload is read, so an anonymous one is refused without its
// file ever being held in memory - the validator below is what reads it.
app.post('/api/images', authenticate);

/**
 * Every request is checked against the spec before it reaches a route, so the spec is the one place
 * that says what a well formed request is. Who may do what is still decided by the routes, and the
 * token by `authenticate` - the spec only says that one is needed.
 */
app.use(OpenApiValidator.middleware({
    // The validator resolves the spec in place, and Swagger UI is still serving the same object
    apiSpec: structuredClone(apiDocs) as OpenAPIV3.DocumentV3,
    validateRequests: { coerceTypes: true },
    validateResponses: false,
    validateSecurity: false,
    fileUploader: {
        storage: multer.memoryStorage(),
        limits: { fileSize: config.maxUploadBytes, files: 1 }
    }
}));

/** What someone surfing to the bare address sees - what this is, who it is for and where its code is */
app.get('/', (req: Request, res: Response) => {
    res.set('Cache-Control', 'public, max-age=3600');
    res.type('html').send(renderHomePage(config.publicBaseUrl, {
        thumbnailWidths: config.thumbnailWidths,
        faviconUrl: config.faviconUrl
    }));
});

app.get('/health', (req: Request, res: Response) => {
    res.json({ status: 'ok' });
});

/**
 * A browser asks the root of a site for its icon whenever what it is showing is not a page it can
 * put a `<link rel="icon">` in - which here is every picture and every piece of metadata, since
 * those are what this service mostly serves. Answering it gives the tab an icon for those too.
 */
app.get('/favicon.ico', (req: Request, res: Response) => {
    if (!config.faviconUrl) {
        return res.sendStatus(404);
    }
    res.set('Cache-Control', 'public, max-age=86400');
    res.redirect(301, config.faviconUrl);
});

// --- Upload ---

/** The fields of an upload, as the validator leaves them - checked against the spec and coerced */
type UploadBody = {
    lat?: number;
    lng?: number;
    license?: string;
    osmUser?: string;
    osmUserId?: string;
    uploadedAt?: string;
};

/**
 * The spec has already made each coordinate a number in range. What it can not say is that they only
 * mean something together, so a lone one is refused here.
 */
function parseLocation(body: UploadBody): { lat: number; lng: number } | undefined | 'invalid' {
    if (body.lat === undefined && body.lng === undefined) {
        return undefined;
    }
    if (body.lat === undefined || body.lng === undefined) {
        return 'invalid';
    }
    return { lat: body.lat, lng: body.lng };
}

type Attribution = {
    osmUser: string;
    osmUserId?: string;
    uploadedAt: string;
};

/**
 * Who an image is credited to and when it was contributed.
 *
 * Both are taken from the access token and the clock, so an image can not be attributed to someone
 * who did not send it. A moderator is the one exception, and only so that an archive that already
 * exists somewhere else can be brought in with the people who actually took the pictures still on
 * them - which is the whole point of moving it. What a moderator states is not verified.
 *
 * `osmUserId` may be left out there, and only there, because an account that has since been deleted
 * has no id left to give - the name is still who took the picture and still who to credit. No token
 * can match an image with no id on it, so only a moderator can ever remove one.
 *
 * `osmUser` may be left out too, and then the moderator is credited as usual. An archive holds
 * pictures whose author it never recorded, and the date is still worth keeping even though there is
 * nobody but the importer to put on them.
 */
function parseAttribution(body: UploadBody, user: AuthenticatedUser): Attribution | 'invalid' | 'forbidden' {
    const stated = body.osmUser || body.osmUserId || body.uploadedAt;
    if (!stated) {
        return { osmUser: user.osmUser, osmUserId: user.osmUserId, uploadedAt: new Date().toISOString() };
    }
    if (!config.adminOsmUserIds.includes(user.osmUserId)) {
        return 'forbidden';
    }
    if (body.osmUserId && !body.osmUser) {
        return 'invalid';
    }
    return {
        osmUser: body.osmUser || user.osmUser,
        osmUserId: body.osmUser ? (body.osmUserId || undefined) : user.osmUserId,
        uploadedAt: body.uploadedAt || new Date().toISOString()
    };
}

app.post('/api/images', async (req: Request, res: Response, next: NextFunction) => {
    try {
        // The spec requires the file and allows only one, so the validator has already made sure it is there
        const file = (req.files as Express.Multer.File[]).find(uploaded => uploaded.fieldname === 'file')!;
        const probed = await probe(file.buffer);
        if (!probed) {
            return res.status(400).json({ message: 'The file is not a jpeg, a png or a webp' });
        }
        const body = req.body as UploadBody;
        const location = parseLocation(body);
        if (location === 'invalid') {
            return res.status(400).json({ message: 'lat and lng must be sent together' });
        }
        // The spec only lets a permissive license through, and the default is checked when the service starts
        const license = body.license || config.defaultLicense;
        const attribution = parseAttribution(body, req.user!);
        if (attribution === 'forbidden') {
            return res.status(403).json({ message: 'Only a moderator can state who an image belongs to' });
        }
        if (attribution === 'invalid') {
            return res.status(400).json({ message: 'osmUserId needs the osmUser it belongs to' });
        }

        const id = idOf(file.buffer);
        const existingMetadata = await readMetadata(id);
        if (existingMetadata) {
            return res.status(200).json(existingMetadata);
        }

        const metadata: ImageMetadata = {
            id,
            url: urlOf(id, probed.format),
            osmUser: attribution.osmUser,
            osmUserId: attribution.osmUserId,
            location,
            license,
            uploadedAt: attribution.uploadedAt
        };
        await saveImage(file.buffer, metadata);
        res.status(201).json(metadata);
    } catch (error) {
        next(error);
    }
});

app.delete(new RegExp(`^/api/images/(${ID_PATTERN})$`), authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        const id = req.params[0] as string;
        const metadata = await readMetadata(id);
        if (!metadata) {
            return res.sendStatus(404);
        }
        const user = req.user!;
        if (metadata.osmUserId !== user.osmUserId && !config.adminOsmUserIds.includes(user.osmUserId)) {
            return res.status(403).json({ message: 'Only the uploader can delete an image' });
        }
        await deleteImage(metadata);
        res.sendStatus(204);
    } catch (error) {
        next(error);
    }
});

// --- Serving ---

function sendStoredFile(res: Response, filePath: string, next: NextFunction) {
    res.sendFile(filePath, { headers: { 'Cache-Control': IMMUTABLE_CACHE_CONTROL } }, error => {
        if (!error) {
            return;
        }
        if ((error as NodeJS.ErrnoException & { status?: number }).status === 404) {
            return res.sendStatus(404);
        }
        next(error);
    });
}

app.get(METADATA_ROUTE, async (req: Request, res: Response, next: NextFunction) => {
    try {
        const metadata = await readMetadata(req.params[0] as string);
        if (!metadata) {
            return res.sendStatus(404);
        }
        res.set('Cache-Control', IMMUTABLE_CACHE_CONTROL);
        res.json(metadata);
    } catch (error) {
        next(error);
    }
});

/**
 * Thumbnails are generated the first time they are asked for and then kept on disk, so a width that
 * nobody uses costs nothing. Requests that arrive together for a thumbnail that does not exist yet
 * share a single generation rather than each doing the same work.
 */
const thumbnailsBeingCreated = new Map<string, Promise<void>>();

async function ensureThumbnail(id: string, format: ImageFormat, width: number): Promise<void> {
    const targetPath = thumbnailPath(id, width, format);
    if (await exists(targetPath)) {
        return;
    }
    const inFlight = thumbnailsBeingCreated.get(targetPath);
    if (inFlight) {
        return inFlight;
    }
    const creation = (async () => {
        const sourcePath = originalPath(id, format);
        // A thumbnail at least as wide as the picture is the picture - resizing would only re-encode
        // it into a second copy of the same thing. It gets a name under `thumb` all the same, so the
        // front end keeps serving every size straight from disk, but the bytes are stored once.
        // The size is read off the header rather than by decoding the whole picture.
        const { width: originalWidth } = await measure(sourcePath);
        if (originalWidth !== undefined && width >= originalWidth) {
            return linkFileAtomically(sourcePath, targetPath);
        }
        const original = await readFile(sourcePath);
        const thumbnail = await createThumbnail(original, width);
        await writeFileAtomically(targetPath, thumbnail);
    })().finally(() => thumbnailsBeingCreated.delete(targetPath));
    thumbnailsBeingCreated.set(targetPath, creation);
    return creation;
}

/**
 * Serves a picture, at its own size or at one of the widths this instance generates.
 *
 * The width is a query parameter rather than a path of its own, so that every size of a picture is
 * the same url with one thing varied - whoever holds `<id>.jpg` asks for a smaller one by appending
 * `?width=`, without having to know how to spell a second kind of url. Only the widths the instance
 * allows are generated, since an arbitrary width would let anyone fill the disk with derivatives.
 */
/**
 * The page a credit under an image links to, so that whoever follows it sees the picture together
 * with who took it and the license it may be used under.
 */
app.get(PAGE_ROUTE, async (req: Request, res: Response, next: NextFunction) => {
    try {
        const id = req.params[0] as string;
        const metadata = await readMetadata(id);
        if (!metadata) {
            return res.sendStatus(404);
        }
        // The size and the camera come out of the picture as the page is built. They are not stored
        // with the rest, since the picture is the thing that knows them and it can never disagree
        // with itself the way a copy of them could.
        const format = formatOf(metadata);
        const measured = await measure(originalPath(id, format));
        res.type('html').send(renderImagePage(metadata, format, measured, {
            thumbnailWidths: config.thumbnailWidths,
            faviconUrl: config.faviconUrl
        }));
    } catch (error) {
        next(error);
    }
});

app.get(IMAGE_ROUTE, async (req: Request, res: Response, next: NextFunction) => {
    try {
        const id = req.params[0] as string;
        const format = req.params[1] as ImageFormat;
        const requestedWidth = req.query.width;

        if (requestedWidth === undefined) {
            if (!await exists(originalPath(id, format))) {
                return res.sendStatus(404);
            }
            return sendStoredFile(res, originalPath(id, format), next);
        }

        const width = Number(requestedWidth);
        if (!config.thumbnailWidths.includes(width)) {
            return res.status(400).json({
                message: 'Unsupported width',
                supportedWidths: config.thumbnailWidths
            });
        }
        if (!await exists(originalPath(id, format))) {
            return res.sendStatus(404);
        }
        await ensureThumbnail(id, format, width);
        sendStoredFile(res, thumbnailPath(id, width, format), next);
    } catch (error) {
        next(error);
    }
});

// --- Error Handling ---

app.use((err: any, req: Request, res: Response, next: NextFunction) => {
    // The validator reads the upload, and turns a file over the limit into a 413 of its own
    if (err?.status === 413) {
        return res.status(413).json({ message: `The image is larger than ${config.maxUploadBytes} bytes` });
    }
    res.status(err.status || 500).json({
        message: err.message,
        errors: err.errors,
    });
    console.log(req.method + ' ' + req.url + ' : ' + err.message);
});