# user-images.openapi.yml
openapi: 3.0.1
info:
  title: Mapeak User Images
  description: |
    Stores the images users attach to points of interest, and serves them at a stable url that an
    OSM entity can hold in its `image` tag.

    An image is addressed by the hash of its bytes, so the same image uploaded twice gets the same
    url, and the bytes behind a url never change.
  version: 1.0.0
servers:
  - url: https://images.mapeak.com
    description: Production
  - url: http://localhost:3000
    description: Local
paths:
...
  /favicon.ico:
    get:
      tags:
        - Pages
      summary: Redirects to the icon of the site this instance belongs to
      responses:
        '301':
          description: The icon
        '404':
          description: The instance has no icon
...
  /api/images:
    post:
      tags:
        - Images
      summary: Uploads an image
      description: |
        The uploader is taken from the OSM access token and never from the request, so an image is
        always credited to the account that actually sent it.

        A moderator of the instance is the one exception: it may state `osmUser`, `osmUserId` and
        `uploadedAt`, so that an archive that already exists somewhere else can be brought in with
        the people who actually took the pictures still on them. Anyone else sending those fields is
        refused.

        Uploading the same bytes again is not an error, it returns the existing image with a 200.
      security:
        - osmAccessToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The image itself, a jpeg, a png or a webp
                lat:
                  type: number
                  format: double
                  minimum: -90
                  maximum: 90
                  description: The latitude the image belongs to, sent together with lng
                lng:
                  type: number
                  format: double
                  minimum: -180
                  maximum: 180
                  description: The longitude the image belongs to, sent together with lat
                license:
                  type: string
                  enum: [CC0-1.0, CC-BY-4.0, CC-BY-3.0]
                  description: |
                    The SPDX id of the license, defaults to the one of the instance. Only permissive
                    licenses are accepted, so that whatever shows a picture is free to.
                osmUser:
                  type: string
                  description: |
                    Moderators only. Who to credit the image to. Defaults to the account the
                    access token belongs to, which is what an archive that never recorded an author
                    still needs when its date is worth keeping.
                osmUserId:
                  type: string
                  description: |
                    Moderators only. May be left out when the OSM account no longer exists, and then
                    the image is credited by name and only a moderator can remove it
                uploadedAt:
                  type: string
                  format: date-time
                  description: |
                    Moderators only. When the image was originally contributed, as opposed to when
                    it reached this instance. Defaults to now
      responses:
        '201':
          description: The image was stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageMetadata'
        '200':
          description: The same image was already stored, its existing metadata is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageMetadata'
        '400':
          description: The file is missing, is not an image, or a field is malformed
        '401':
          description: The OSM access token is missing or invalid
        '403':
          description: The caller is not a moderator and tried to state who the image belongs to
        '413':
          description: The image is larger than the limit of the instance
...
components:
...
  securitySchemes:
    osmAccessToken:
      type: http
      scheme: bearer
      description: An OpenStreetMap OAuth2 access token, the one the site holds for the logged in user
  schemas:
    ImageMetadata:
      type: object
      required:
        - id
        - url
        - osmUser
        - license
        - uploadedAt
      properties:
        id:
          type: string
          example: 3a7f9c2e1b4d0a95f60c2d3e4f501b6c
        url:
          type: string
          description: The url to put in the OSM "image" tag
          example: https://images.mapeak.com/3a7f9c2e1b4d0a95f60c2d3e4f501b6c.jpg
        osmUser:
          type: string
          description: The OSM display name of whoever uploaded the image, who should be credited for it
        osmUserId:
          type: string
        location:
          type: object
          properties:
            lat:
              type: number
              format: double
            lng:
              type: number
              format: double
        license:
          type: string
          example: CC0-1.0
        uploadedAt:
          type: string
          format: date-time